apps_reworked/osint_rework/requirements.txt
fastapi
uvicorn
pywebview
requests
dnspython
python-whois
ipwhois
